pyproject.toml
[project]
name = "finai-credit-os"
version = "0.1.0"
description = "AI-Native Lending Operations Platform — GrāmCredit AI"
requires-python = ">=3.10"
dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",
    # Agent framework
    "langgraph>=0.4.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    # LLM
    "anthropic>=0.40.0",
    # Data
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    # Observability
    "langfuse>=2.40.0",
    "structlog>=24.1.0",
    # Memory
    "mem0ai>=0.1.0",
    # Utils
    "python-dateutil>=2.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0",
    "types-python-dateutil>=2.9.0",
    "types-redis>=4.6.0",
]
demo = [
    "reportlab>=4.2.0",
    "faker>=25.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/finai"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["finai"]

[tool.mypy]
python_version = "3.10"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["tests/", "demo/"]

[[tool.mypy.overrides]]
module = [
    "mem0ai.*",
    "langfuse.*",
    "langgraph.*",
    "langgraph_checkpoint_postgres.*",
    "langchain_core.*",
    "langchain_anthropic.*",
    "celery.*",
    "reportlab.*",
    "faker.*",
    "composio.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
